+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Installation
 Pragmas and options
 -------------------
 
+* New option `--print-options` to print a simple list of all options.
+  This list can e.g. be used to implement bash completion.
+
 Warnings
 --------